Always fail if IPv6 socket cannot be created
IPv6 socket creation is not fatal unless a v6 src_ip has been passed,
but that is an optional value that is not tied to the server address,
which can still be IPv6 if no src_ip was passed at all.
In the talk_radius() function, in case the server address is v6, the
v6 socket is selected regardless of the src_ip.