Use per-server socket if src_ip is set in config file
The config file might specify a different src_ip for each server
address. In order to support it, create a socket per server if the
src_ip is set for it. As an optimisation, to avoid always creating
and binding many sockets when it's not necessary, only open a
per-server socket if the src_ip is set for that server, and use the
global conf socket otherwise.